I init <Checkboxes /> component.
